requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
marshmallow
python-dotenv
SQLAlchemy
